Fix dtype in static cache example to prevent buffer mismatch
Add dtype='f32' to the build() call in static_cache_generation.py,
matching the convention in text_generation.py. This ensures the model
weights are f32, consistent with the np.float32 cache buffers.

Agent-signed-off: Developer (a99db0a1) [claude-opus-4.6]

+++ examples/static_cache_generation.py
@@ -59,7 +59,7 @@ MAX_SEQ_LEN = 2048
 def build_model(model_id: str, *, max_seq_len: int):
     """Build an ONNX model with static KV cache."""
     task = CausalLMTask(static_cache=True, max_seq_len=max_seq_len)
-    return build(model_id, task=task)
+    return build(model_id, task=task, dtype="f32")
 
 
 # ---------------------------------------------------------------------------
